Add UpdateAvaliabilityTime endpoint to controller
This commit introduces a new HTTP PUT method, `UpdateAvaliabilityTime`, in the `VenueAvaliabilityController.cs`. It allows updating the availability time for a venue by accepting an `id` as a query parameter and a `UpdateVenueAvaliabilityDto` in the request body. Additionally, new using directives for `venue_service.Src.Dtos` and `venue_service.Src.Services` are added to facilitate access to necessary data transfer objects and services.