Add availability filter to venue query
This update introduces a new `Where` clause in the
`VenueService.cs` file that filters venues based on
their availability times. It ensures that only venues
with availability times having an `EndDate` greater
than or equal to the current UTC date and time are
returned, enhancing the query to reflect only currently
available venues.